workspaces: Allow non-existing workspaces
When using quickswitch with workspaces, it's very helpful to be able to
create new ones on the fly using the same i3 bindings as you would for
when dealing with pre-existing ones.

When no target is found in the lookup table, this patch overrides the
lookup result and simply uses whatever workspace name the user supplied.

In my view, this behaviour is much more intuitive, and tremendously
helpful.